seperating linsearch
